Try to fix broken actions.
The 3.8/macOS python-package job failed due to a scipy package with a
wrong/suspicious hash.  It was trying to install 1.4, which is less than the
minimum value of 1.7 specified for the package.  Trying 1.7 to see if it passes
again.  Sphinx package builds failed with a strange error.  I find in stack
overflow that this error message can be related with shallow git clones.
Asking action to download the full repo as suggested to see if that fixes it.