(Issue #106) Inject an "error"
I had a very difficult time trying to inject a true error.  I was able to get
the sphinx-build output to say that it found "errors" by writing "tx.rst"
instead of "tox.rst" when including that file and by writing ":mth:" instead of
":math:".  However, neither of these actually resulted in a non-zero exit code
unless I used the treat warnings as errors flag.  In this case, the
sphinx-build output says that it is treating a warning as an error.  Therefore,
I would *never* build with sphinx without this flag in an automated scheme
since it appears that sphinx doesn't actually have errors.