catchup with Develop (#116)
* Update README.rst

* Add logos via upload

* resize logos

* include optional dependencies in pyproject.toml dynamic field

* eliminating duplicate info in setup.py and pyproject.toml

* move scikit-learn to optional dependency and update installation guide.

* node20 github action specification

* remove PTMC, ptemcee dependent sampler

* First draft of tox integration.

Was able to run all tox tasks successfully.  Did quick review of coverage
reports and sphinx output.  All looked acceptable.  Flake8 shows all passing,
which is as to be expected since it is already integrated in the repo.  Sphinx
output printed a few warnings.

* Add tox information to developer's guide.

Built the HTML form of the docs and all looks good.  The language should be
cleaned up during PR review.

* Clean-up sphinx documentation to get rid of an error and a warning.

* Add sphinx docs build GitHub action

* Try to fix broken actions.

The 3.8/macOS python-package job failed due to a scipy package with a
wrong/suspicious hash.  It was trying to install 1.4, which is less than the
minimum value of 1.7 specified for the package.  Trying 1.7 to see if it passes
again.  Sphinx package builds failed with a strange error.  I find in stack
overflow that this error message can be related with shallow git clones.
Asking action to download the full repo as suggested to see if that fixes it.

* Still trying to fix actions.

* Both actions pass, but are setup to use deprecated Node.js 16 nodes.

Trying to see if updating to latest version of GH action tools works.

* Update setup-python to overcome deprecation issue.

* Hopefully this will do it.

* Try code coverage with tox.  Not pushing results to a server.

* remove py38 from workflow

* (Issue #106) Try pushing to coveralls.

* (Issue #106) Cleaner apt-get usage.

* (Issue #112) Officially remove 3.8 support & clean.

* (Issue #106) Use tox flake8 task in own action.

* (Issue #106) Remove flake8 from test action.

* (Issue #106) Cleaning build docs action

* (Issue #106) Need explicit installation of LaTeX.

* (Issue #106) Continue cleaning actions.

* More conservative RTD settings.

Currently HTML and PDF forms of Sphinx docs build without errors and warnings.
To keep the docs in the high-level of cleanliness, we ask RTD to treat warnings
as errors.  I'm presently not using this setting with tox so that users get the
normal experience there.

* (Issue #106) Cleaning as part of PR review.

* (Issue #106) Remove requirements.txt.

Dependencies are already encoded in pyproject.toml.

* (Issue #106) Add Coveralls badge.

* (Issue #106) Update README to reflect changes made in branch.

* (Issue #106) remove the 3

* remove dependency specifications, point to pyproject.toml

* (Issue #106) Refer documentation builds to tox in developers guide.

* (Issue #106) Intentionally violate PEP8.

Flake8 should complain about importing but not using os and not having enough
blank lines.  THIS SHOULD BE REVERTED!

* Revert "(Issue #106) Intentionally violate PEP8."

This reverts commit 3b239c22113697f224fd28b1df1298f321347fc4.

* (Issue #106) Intentionally break a test.

A single test should fail because it incorrectly expects TypeError instead of
ValueError.  THIS SHOULD BE REVERTED!

* Revert "(Issue #106) Intentionally break a test."

This reverts commit 18128a0d4ad25a72ba717ac4853a72ad9dcc7f79.

* (Issue #106) Change docs to inject a warning.

We should see the warning treated as an error by CI.  THIS SHOULD BE REVERTED!

* Revert "(Issue #106) Change docs to inject a warning."

This reverts commit ab07daf068fd6a223b9ebace1ca170a7de5c5481.

* (Issue #2) Intentionally inject error in docs.

THIS SHOULD BE REVERTED!

* (Issue #106) Reestablish warnings as errors and clean docs folder.

* (Issue #106) Inject an "error"

I had a very difficult time trying to inject a true error.  I was able to get
the sphinx-build output to say that it found "errors" by writing "tx.rst"
instead of "tox.rst" when including that file and by writing ":mth:" instead of
":math:".  However, neither of these actually resulted in a non-zero exit code
unless I used the treat warnings as errors flag.  In this case, the
sphinx-build output says that it is treating a warning as an error.  Therefore,
I would *never* build with sphinx without this flag in an automated scheme
since it appears that sphinx doesn't actually have errors.

* (Issue #106) Revert intentional documentation break.

* (Issue #106) Add maintenance reminders.

* (Issue #106) Inject untested lines to see change in coverage.

* Revert "(Issue #106) Inject untested lines to see change in coverage."

This reverts commit be200d5b63492202b79823789ef89cac7465a006.

* (Issue #106) Clean tox docs.

* (Issue #106) See if tests pass with 3.12.

* Use latest versions of dependencies

* Put upper limit on Python version to see 3.12 failures in CI.

* (Issue #106) Revert pyproject.toml to nominal setup.

* Use 3.12 for all single-version CI actions.

* Sphinx build failing for 3.12.

* excluding venvs and local build from flake8

---------

Co-authored-by: Stefan M Wild <wild@lbl.gov>
Co-authored-by: Jared O'Neal <joneal@anl.gov>