(Issue #134) Try running a "full" test suite outside tox.
Since the tests are not yet in the Python package, running a full test suite via
tox is challenging.  Since we plan to integrate the tests into the package soon,
we prefer to do the testing at a lower level here for now.

When I run this "full" test suite locally, some tests are failing.

TODO: Moses will have to build out the test suite until it is actually full and
get all tests passing.