(Issue #137) Store Cython-generated .c file in repo.
This is recommended by both the Cython and setuptools docs.  As a result Cython
need only be installed by developers that are actually working on that file.
setup.py is responsible only for compiling the c file, which requires numpy
headers but not cython.  This can be seen in the changes here including the fact
that the test suite no longer needs to cythonize the c file to run.

I added a cythonize tox task to setup a venv with Cython and generate the c file
with it.  The coverage test is still passing.  I was also able to install
surmise in a venv and run the GP_surmise notebook sucessfully, which should be
using the Cython code.