(Issue #137) Continue cleaning C code management scheme.
The inability that I had with installing numpy < 1.26 in build systems is no
longer an issue since we are now fixed at the more modern 2.0.0.  On my system,
I was able to get tests working with 1.22.0.

Developers should all use the same Cython version.

Explicitly fix the Cython language directive to Python 3 to get rid of warnings.
According to the warning, it was previously using 3str by default.  Therefore,
there are a few changes to the .c file, which was generated using the new Cython
version and the target 2.0.0 numpy version.