(Issue #74) Expanding testing capabilities.
If users want to run just a subset of the tests, they probably want to be in
editable/developer mode.  This decision is likely only consequential in that it
might make the tasks run a bit faster and if developers decide to activate one
of the associated venvs instead of coverage or nocoverage.