(Issue #74) Don't provide flags for user to specify which method to test.
From earlier testing, I don't think that those flags are used anymore.  Rather,
I think that pytest ignores them and always runs the test setups specified with
the parameterize decorator.  Moses doesn't think that this level of fine-grained
test control is needed for now.  We could try to add it in later when there is a
genuine need.

As part of this, I reverted variable names so that the PR review should be
easier.