(Issue #74) Make scikit-learn a required dependence.
Since it is required for running the test suite and test() is now part of the
package, Moses thought that it made sense to make this change.  While we are
adding a new dependence, it is a mature, well-known, and widely-used package
that most surmise users would be familiar with.  In addition, this is a
significant simplification in terms of code, infrastucture, and documentation.

All tox tasks are running successfully and the docs look good.