(Issue #170) First draft of new Ball Drop example notebook.
This should be a cleaner version of the previous examples 1 and 2 (now removed).
The target audience is graduate students who are not necessarily physicists and
that are just starting out.

The notebook runs through and all content looks good.  I could generate the book
without issue and the rendered content looks good as well.  There are some
simple mistakes and typos that need to be fixed.